requirements.txt
python-telegram-bot
torchvision
transformers
torch